Add generate and download gui feature

+++ DFD2GUI/routes.py
@@ -10,16 +10,21 @@ import os
 import shutil
 import json
 
+sample_data = {
+    'int': '256',
+    'string': 'Boulevard st',
+    'float': '12.5',
+    'date': '2020-12-31',
+    'boolean': 'Yes'
+}
 
 def activate_link(page):
     active_link = {'dashboard': '', 'project-list': '', 'new-project': ''}
     active_link[page] = 'active'
     return active_link
 
-
 def reset_project_session():
     project_session = {'project_name': '', 'curr_route': '', 'path': ''}
-
 
 @app.route("/", methods=["POST", "GET"])
 @app.route("/login",  methods=['POST', 'GET'])
@@ -37,7 +42,6 @@ def login():
         else:
             flash("Your email or password is wrong", "danger")
     return render_template('login.html', title="Login", form=form)
-
 
 @app.route("/register", methods=['POST', 'GET'])
 def register():
@@ -62,7 +66,6 @@ def register():
         return redirect(url_for('login'))
     return render_template('register.html', title="Register", form=form)
 
-
 @app.route("/dashboard")
 @login_required
 def dashboard():
@@ -77,7 +80,6 @@ def dashboard():
 
     return render_template('dashboard.html', title="Dashboard", active_link=activate_link('dashboard'), projects=projects)
 
-
 @app.route("/project-list")
 @login_required
 def project_list():
@@ -85,7 +87,6 @@ def project_list():
         user_id=current_user.id).order_by(Project.id)
     project_list = list(enumerate(project_list, 1))
     return render_template('project_list.html', title="Project list", active_link=activate_link('project-list'), projects=project_list)
-
 
 @app.route("/new-project", methods=["POST", "GET"])
 @login_required
@@ -111,7 +112,6 @@ def new_project():
         print(form.errors)
     return render_template('new_project.html', title="New Project", form=form, active_link=activate_link('new-project'))
 
-
 @app.route("/add-entity", methods=["POST", "GET"])
 @login_required
 def add_entity():
@@ -119,7 +119,6 @@ def add_entity():
     project_session['curr_route'] = 'add_entity'
     print(project_session)
     return render_template('add_entity.html', title="Add Entity", active_link=activate_link('new-project'))
-
 
 @app.route("/add-entity-func", methods=["POST", "GET"])
 @login_required
@@ -136,12 +135,10 @@ def add_entity_func():
         f.write(json_entity)
     return redirect(url_for("add_datastore"))
 
-
 @app.route("/add-datastore", methods=["POST", "GET"])
 @login_required
 def add_datastore():
     return render_template('add_datastore.html', title="Add Datastore", active_link=activate_link('new-project'))
-
 
 @app.route("/add-datastore-func")
 @login_required
@@ -159,12 +156,10 @@ def add_datastore_func():
         f.write(json_datastore)
     return redirect(url_for("add_process"))
 
-
 @app.route("/add-process", methods=["POST", "GET"])
 @login_required
 def add_process():
     return render_template('add_process.html', title="Add Process", active_link=activate_link('new-project'))
-
 
 @app.route("/add-process-func")
 @login_required
@@ -185,7 +180,6 @@ def add_process_func():
         f.write(process_json)
     return redirect(url_for("add_process_det"))
 
-
 @app.route("/add-process-det")
 @login_required
 def add_process_det():
@@ -201,7 +195,6 @@ def add_process_det():
                       'name': dic_process[key]['name'], 'parent': dic_process[key]['parent']}}
         lowest_process_list.append(temp)
     return render_template('add_process_det.html', title="Add Process Detail", active_link=activate_link('new-project'), process=lowest_process_list)
-
 
 @app.route('/add-process-det-func')
 @login_required
@@ -222,7 +215,6 @@ def add_process_det_func():
 
     return redirect(url_for('add_relation'))
 
-
 @app.route("/add-relation", methods=["POST", "GET"])
 @login_required
 def add_relation():
@@ -234,7 +226,6 @@ def add_relation():
     out_dic = get_rel_input_data(in_dic)
     test = json.dumps(out_dic)
     return render_template('add_relation.html', title="Add relation", active_link=activate_link('new-project'), in_data=test)
-
 
 @app.route("/add-relation-func", methods=["POST", "GET"])
 @login_required
@@ -272,7 +263,6 @@ def add_relation_func():
 
     return redirect(url_for('add_gui_attr'))
 
-
 @app.route("/add-gui-attr")
 @login_required
 def add_gui_attr():
@@ -292,7 +282,6 @@ def add_gui_attr():
                         'gui_type': temp_gui_type})
     lis = json.dumps(lis)
     return render_template('add_gui_attr.html', title="Add GUI atrributes", active_link=activate_link('new-project'), data=lis)
-
 
 @app.route("/add-gui-attr-func")
 @login_required
@@ -316,7 +305,6 @@ def add_gui_attr_func():
         f.write(out_json)
 
     return redirect(url_for('dashboard'))
-
 
 @app.route("/project/<int:project_id>", methods=["POST", "GET"])
 @login_required
@@ -402,23 +390,25 @@ def project(project_id):
 # def project_edit(project_id):
 #     return f"<h1>Project id: {project_id} </h1>"
 
-
-@app.route('/generate-gui')
-@login_required
-def generate_gui():
+def generate_gui(project_id):
+    # Check ownership
+    project_info = Project.query.filter_by(id=project_id).first()
+    if current_user.id != project_info.user_id:
+        return redirect(url_for('dashboard'))
+    # Delete existing generated GUI
     from shutil import copytree
-    path = os.path.join(app.root_path, 'project_output sample')
-    os.chdir(path)
-    for i in os.listdir(path):
+    path = os.path.join(app.root_path, 'user_project',current_user.email, project_info.project_name)
+    os.chdir(os.path.join(path, 'GUI'))
+    for i in os.listdir(os.path.join(path,'GUI')):
         if os.path.isdir(i):
             shutil.rmtree(i)
         else:
             os.remove(i)
     src = os.path.join(app.root_path, 'gui_template', 'assets')
-    dst = os.path.join(app.root_path, 'project_output sample', 'assets')
+    dst = os.path.join(path, 'GUI', 'assets')
     copytree(src, dst)
     os.chdir(app.root_path)
-
+    # Generate sidebar nav link
     side_bar_link = '''
     <li class="nav-item">
         <a href="{link}.html" class="nav-link">
@@ -427,11 +417,12 @@ def generate_gui():
     </li>
     '''
     # get metadata
-    with open(os.path.join(app.root_path, 'test.json')) as f:
+    with open(os.path.join(path,'metadata.json'), 'r') as f:
         txt = f.read()
     json_dic = json.loads(txt)
     sidebar = []
     # get html template
+    html_output = {}
     with open(os.path.join(app.root_path, 'gui_template', 'template.html'), 'r') as f:
         html = f.read()
     # get gui process
@@ -444,19 +435,47 @@ def generate_gui():
         proc_name = i['name']
         link = proc_name.replace(' ', '')
         sidebar.append(side_bar_link.format(link=link, process_name=proc_name))
-    html = html.format(sidebar=''.join(sidebar), project_name='Test 1', process_name='{process_name}', gui_type='{gui_type}')
+    html = html.format(sidebar=''.join(sidebar), project_name='Test 1', process_name='{process_name}', gui_type='{gui_type}', content='{content}')
     # Generate html page for every process gui
     for i in process_gui:
         link = f"{i['name'].replace(' ','')}.html"
         path = os.path.join(app.root_path, 'project_output sample', link)
-        temp_template = html.format(process_name=i['name'], gui_type= gui_type_name_tranf(i['gui']))
+        temp_template = html.format(process_name=i['name'], gui_type= gui_type_name_tranf(i['gui']), content='{content}')
+        html_output[i['name']] = temp_template
+    # Generate Content
+    for i in process_gui:
+        if i['gui'] == 'master_data':
+            content = create_md_template(i['gui_attr'])
+            html_output[i['name']] = html_output[i['name']].format(content=content)
+        elif i['gui'] == 'grid':
+            content = create_grid_template(i['gui_attr'])
+            html_output[i['name']] = html_output[i['name']].format(content=content)
+        elif i['gui'] == 'input':
+            content = create_input_template(i['gui_attr'])
+            html_output[i['name']] = html_output[i['name']].format(content=content)
+    # Write to HTML
+    for name, html in html_output.items():
+        path = os.path.join(app.root_path, 'user_project',current_user.email, project_info.project_name, 'GUI', '{}.html'.format(name))
         with open(path, 'w') as f:
-            f.write(temp_template)
+            f.write(html)
 
-
-    # with open(os.path.join(app.root_path, 'project_output sample', 'test.html'), 'w') as f:
-    #     f.write(html)
-    return 'worked'
+@app.route('/download-gui/<int:project_id>')
+@login_required
+def download_gui(project_id):
+    # Get project info
+    project_info = Project.query.filter_by(id=project_id).first()
+    # Generate HTML
+    generate_gui(project_id)
+    # Delete existing .zip file if available
+    os.chdir(os.path.join(app.root_path, 'user_project', current_user.email, project_info.project_name))
+    if 'GUI.zip' in os.listdir():
+        os.remove('GUI.zip')
+    # Make archive .zip
+    from shutil import make_archive
+    make_archive('GUI', 'zip', root_dir=os.path.join(os.getcwd(), 'GUI'))
+    return send_from_directory(directory=os.getcwd(), filename='GUI.zip', as_attachment=True)
+    # return os.getcwd()
+    
 
 @app.route('/test')
 def test():
@@ -465,30 +484,31 @@ def test():
 @app.route("/fuck-this-shit")
 def fuck_this_shit():
     user_project = os.path.join(app.root_path, 'user_project')
-    os.chdir(user_project)
-    for i in os.listdir(user_project):
-        if os.path.isdir(i):
-            shutil.rmtree(i)
-        else:
-            os.remove(i)
+    folder = user_project
+    for filename in os.listdir(folder):
+        file_path = os.path.join(folder, filename)
+        try:
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+        except Exception as e:
+            print('Failed to delete %s. Reason: %s' % (file_path, e))
 
     db.drop_all()
     db.create_all()
     return redirect(url_for('logout'))
-
 
 @app.route("/logout")
 def logout():
     logout_user()
     return redirect(url_for('login'))
 
-
 def check_lowest(key, dic):
     for k in dic:
         if dic[k]['parent'] == dic[key]['name']:
             return False
     return True
-
 
 def get_lowest(dic):
     output = []
@@ -500,7 +520,6 @@ def get_lowest(dic):
         if check_lowest(key, temp_process):
             output.append(key)
     return output
-
 
 def get_rel_input_data(input_dic):
     rel_input_key = []
@@ -540,3 +559,198 @@ def gui_type_name_tranf(name):
         return 'Grid'
     elif name == 'input':
         return 'Input'
+
+def create_md_template(attr):
+    # Get Template file
+    template_path = os.path.join(app.root_path, 'gui_template', 'master-data-content.html')
+    with open(template_path, 'r') as f:
+        html = f.read()
+    # output template
+    head_template = '''
+    <th>{column_name}</th>
+    '''
+    body_template = '''
+    <td>{data}</td>
+    '''
+    # Writing to HTML template
+    out_head = ''
+    out_body = ''
+
+    for i in attr:
+        out_head += head_template.format(column_name="{name} ({type})".format(name=i['attr_name'], type=i['attr_type']))
+        out_body += body_template.format(data=sample_data[i['attr_type']])
+
+    html = html.format(column_name=out_head, blank_table=out_body)
+
+    return html
+
+def create_grid_template(attr):
+    # Get Template file
+    template_path = os.path.join(app.root_path, 'gui_template', 'grid-data-content.html')
+    with open(template_path, 'r') as f:
+        html = f.read()
+    # Split into 2 section
+    n_data = len(attr)//2
+    remainder = len(attr) % 2
+    middle_section = n_data + remainder
+    left_section = attr[:middle_section]
+    right_section = attr[middle_section:]
+    # section template
+    section_template = '''
+    <tr>
+        <td>
+            <h4 class="font-weight-light">{key}</h4>
+        </td>
+        <td class="pl-2 pr-3">
+            <h4>:</h4>
+        </td>
+        <td>
+            <h4>{value}</h4>
+        </td>
+    </tr>
+    ''' 
+    # add left section data
+    left_content = ''
+    for i in left_section:
+        key = '{name} ({attr_type})'.format(name=i['attr_name'], attr_type=i['attr_type'])
+        value = sample_data[i['attr_type']]
+        left_content += section_template.format(key=key, value=value)
+    # add right section data
+    right_content = ''
+    for i in right_section:
+        key = '{name} ({attr_type})'.format(name=i['attr_name'], attr_type=i['attr_type'])
+        value = sample_data[i['attr_type']]
+        right_content += section_template.format(key=key, value=value)
+    # Fill data
+    html_output = html.format(left_section=left_content, right_section=right_content)
+    return html_output
+
+def create_input_template(attr):
+    # Get Template
+    template_path = os.path.join(app.root_path, 'gui_template', 'input-data-content.html')
+    with open(template_path, 'r') as f:
+        html = f.read()
+    # Input type template
+    input_type_template = {
+        'text': '''
+        <div class="form-group">
+            <label class="col-form-label">{attr_name} ({attr_type})</label>
+            <input type="text" class="form-control">
+        </div>
+        ''',
+        'password':'''
+        <div class="form-group">
+            <label>{attr_name} ({attr_type})</label>
+            <input type="password" placeholder="Password" class="form-control">
+        </div>
+        ''',
+        'date':'''
+        <div class="form-group">
+            <label>{attr_name} ({attr_type})</label>
+            <div class="form-group">
+                <div class="input-group date" id="{id_attr}" data-target-input="nearest">
+                    <input type="text" class="form-control datetimepicker-input" data-target="#{id_attr}">
+                    <div class="input-group-append" data-target="#{id_attr}" data-toggle="datetimepicker">
+                        <div class="input-group-text"><i class="far fa-calendar-alt"></i></div>
+                    </div>
+                </div>
+            </div>
+        </div>
+        ''',
+        'radio':'''
+        <div class="form-group">
+            <label>{attr_name} ({attr_type})</label>
+            <div>
+                <label class="custom-control custom-radio custom-control-inline">
+                    <input type="radio" name="{attr_name}" checked="" class="custom-control-input"><span
+                        class="custom-control-label">Option 1</span>
+                </label>
+                <label class="custom-control custom-radio custom-control-inline">
+                    <input type="radio" name="{attr_name}" class="custom-control-input"><span
+                        class="custom-control-label">Option 2</span>
+                </label>
+                <label class="custom-control custom-radio custom-control-inline">
+                    <input type="radio" name="{attr_name}" class="custom-control-input"><span
+                        class="custom-control-label">Option 3</span>
+                </label>
+            </div>
+            <div>
+                <label class="custom-control custom-radio custom-control-inline">
+                    <input type="radio" name="{attr_name}" checked="" class="custom-control-input"><span
+                        class="custom-control-label">Option 4</span>
+                </label>
+                <label class="custom-control custom-radio custom-control-inline">
+                    <input type="radio" name="{attr_name}" class="custom-control-input"><span
+                        class="custom-control-label">Option 5</span>
+                </label>
+            </div>
+        </div>
+        ''',
+        'checkbox':'''
+        <div class="form-group">
+            <label>{attr_name} ({attr_type})</label>
+            <div>
+                <label class="custom-control custom-checkbox custom-control-inline">
+                    <input type="checkbox" checked="" class="custom-control-input"><span class="custom-control-label">Option
+                        1</span>
+                </label>
+                <label class="custom-control custom-checkbox custom-control-inline">
+                    <input type="checkbox" class="custom-control-input"><span class="custom-control-label">Option 2</span>
+                </label>
+                <label class="custom-control custom-checkbox custom-control-inline">
+                    <input type="checkbox" class="custom-control-input"><span class="custom-control-label">Option 3</span>
+                </label>
+            </div>
+            <div>
+                <label class="custom-control custom-checkbox custom-control-inline">
+                    <input type="checkbox" checked="" class="custom-control-input"><span class="custom-control-label">Option
+                        4</span>
+                </label>
+                <label class="custom-control custom-checkbox custom-control-inline">
+                    <input type="checkbox" class="custom-control-input"><span class="custom-control-label">Option 5</span>
+                </label>
+            </div>
+        </div>
+        ''',
+        'number':'''
+        <div class="form-group">
+            <label class="col-form-label">{attr_name} ({attr_type})</label>
+            <input type="number" class="form-control" placeholder="Numbers">
+        </div>            
+        ''',
+        'select':'''
+        <div class="form-group">
+            <label for="input-select">{attr_name} ({attr_type})</label>
+            <select class="form-control">
+                <option>Option 1</option>
+                <option>Option 2</option>
+                <option>Option 3</option>
+            </select>
+        </div>
+        '''
+    }
+    script_template = '''
+    jQuery(document).ready(function($) {
+    'use strict';
+        %s
+    });
+    '''
+    # Add data
+    form_output = ''
+    script_output = ''
+    script_if = ''
+    for i in attr:
+        attr_type = i['attr_type']
+        attr_name = i['attr_name']
+        if attr_type != 'date':
+            form_output += input_type_template[attr_type].format(attr_name=attr_name, attr_type=attr_type)
+        else:
+            id_attr = attr_name.replace(' ','-')
+            script = "if ($('%s').length) { $('#%s').datepicker({ format: 'L'});}" % (id_attr, id_attr)
+            script_if += script
+            print('\n'*5, script, '\n'*5)
+            form_output += input_type_template[attr_type].format(attr_name=attr_name, attr_type=attr_type, id_attr=id_attr)
+    # render to template
+    script_output += script_template % (script_if)
+    html_out = html.format(form=form_output, script=script_output)
+    return html_out
